マスタエディタ:アクティブになっているタブによって提出確認書への出力がおかしくなるバグを修正。TADをライトした後もう一度ライトすると失敗するバグを修正。
git-svn-id: file:///Users/lillianskinner/Downloads/platinum/twl/TwlToolsRED@81 7061adef-622a-194b-ae81-725974e89856

+++ build/tools/MasterEditor/MasterEditorTWL/MasterEditorTWL/Form_file.cpp
@@ -56,7 +56,11 @@ System::Boolean Form1::saveRom( System::String ^outfile )
 		// 一時ファイルにSRLを書き出しているのでその一時ファイルから出力ファイルを作成
 		System::String ^tmpsrl = this->getSplitTadTmpFilename();
 		result = this->saveSrl( tmpsrl, outfile );
-		System::IO::File::Delete( tmpsrl );
+		if( result )
+		{
+			System::IO::File::Delete( tmpsrl );
+			this->IsLoadTad = false;			// 出力したSRLが次のソースとなる
+		}
 	}
 	else
 	{
